requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
google-generativeai>=0.5.0
Pillow>=10.0.0
python-dotenv>=1.0.0
easyocr>=1.7.0
